Add Spack builds and other updates (#27)
* fixed efs mount point, debugging efa
* efa-install minor fix
* Update impi and efa driver versions
* terraform updates
* impi fix, auto public IP subnet fix
* fix impi modulefile
* mv files from ptmp instead of cp
* can get sg ids now
* safer scrub - was getting unintential deletions
* cleanup after ciofs testing
* added some misc files
* cleanup after intel compiler testing
* re-adding mysettings.tfvars file
* fix region setting
* initial spack installations
* finish Spack builds

+++ cloudflow/workflows/scripts/funcs-setup-instance.sh
@@ -1,11 +1,23 @@
 #!/bin/bash
 
+GCC_VER=8.5.0
+INTEL_VER=2021.3.0
+
+SPACK_DIR=/save/environments/spack
+SPACK_CACHEONLY=0
+
+SPACKOPTS=''
+if [ $SPACK_CACHEONLY -eq 1 ]; then
+  SPACKOPTS='--cache-only'
+fi
+
 # This script will setup the required system components, libraries
 # and tools needed for ROMS forecast models on CentOS 7
 
 #__copyright__ = "Copyright © 2020 RPS Group, Inc. All rights reserved."
 #__license__ = "See LICENSE.txt"
 #__email__ = "patrick.tripp@rpsgroup.com"
+
 
 setup_environment () {
 
@@ -14,7 +26,7 @@ setup_environment () {
   # By default, centos 7 does not install the docs (man pages) for packages, remove that setting here
   sudo sed -i 's/tsflags=nodocs/# &/' /etc/yum.conf
 
-  sudo yum -y update
+  #sudo yum -y update
   sudo yum -y install epel-release
   sudo yum -y install tcsh
   sudo yum -y install ksh
@@ -22,6 +34,10 @@ setup_environment () {
   sudo yum -y install unzip
   sudo yum -y install time.x86_64
   sudo yum -y install glibc-devel
+  sudo yum -y install gcc-c++
+  sudo yum -y install patch
+  sudo yum -y install bzip2
+  sudo yum -y install bzip2-devel
   sudo yum -y install automake
   sudo yum -y install vim-enhanced
   sudo yum -y install environment-modules
@@ -29,7 +45,6 @@ setup_environment () {
 
   cliver="2.2.10"
   curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64-${cliver}.zip" -o "awscliv2.zip"
-  #curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
   /usr/bin/unzip awscliv2.zip
   sudo ./aws/install
   rm awscliv2.zip
@@ -69,11 +84,11 @@ setup_paths () {
     sudo mkdir save
 
     sudo chgrp wheel ptmp
-    sudo chmod 775 ptmp
+    sudo chmod 777 ptmp
     sudo chgrp wheel com
-    sudo chmod 775 com
+    sudo chmod 777 com
     sudo chgrp wheel save
-    sudo chmod 775 save
+    sudo chmod 777 save
   fi
 
   sudo ln -s /mnt/efs/fs1/ptmp /ptmp
@@ -140,11 +155,10 @@ install_efa_driver() {
     sudo mv /usr/lib/modules/$oldkrnl /usr/lib/oldkernel
   done
 
-  # This will get upgraded when we install gcc 6.5
   # Default version is needed to build the kernel driver
   # If gcc has already been upgraded, this will likely fail
   # Should uninstall newer one and install the default 4.8
-  sudo yum -y install gcc
+  # sudo yum -y install gcc
 
   curl -O https://s3-us-west-2.amazonaws.com/aws-efa-installer/$tarfile
   tar -xvf $tarfile
@@ -153,15 +167,106 @@ install_efa_driver() {
   cd aws-efa-installer
 
   # Install without AWS libfabric and OpenMPI, we will use Intel libfabric and MPI
-  sudo ./efa_installer.sh -y --minimal
+  sudo ./efa_installer.sh -y 
 
   # Put old kernels back in original location
-  sudo mv /usr/lib/oldkernel/*  /usr/lib/modules
-  sudo rmdir /usr/lib/oldkernel
+  if [ $(ls /usr/lib/oldkernel/ | wc -l) -ne 0 ]; then
+    sudo mv /usr/lib/oldkernel/*  /usr/lib/modules
+    sudo rmdir /usr/lib/oldkernel
+  fi
 
   cd $home
 }
 
+
+install_spack() {
+
+  home=$PWD
+
+  if [ ! -d /save ] ; then
+    echo "/save does not exst. Setup the paths first."
+    return
+  fi
+
+  mkdir -p $SPACK_DIR
+  git clone https://github.com/spack/spack.git $SPACK_DIR
+  cd $SPACK_DIR
+  git checkout releases/v0.17
+  echo ". $SPACK_DIR/share/spack/setup-env.sh" >> ~/.bashrc
+  echo "source $SPACK_DIR/share/spack/setup-env.csh" >> ~/.tcshrc 
+
+  spack mirror add s3-mirror s3://ioos-cloud-sandbox/public/spack/mirror 
+  spack buildcache update-index -d s3://ioos-cloud-sandbox/public/spack/mirror/
+
+  cd $home
+}
+
+
+install_gcc () {
+
+  home=$PWD
+
+  . $SPACK_DIR/share/spack/setup-env.sh
+
+  #spack install $SPACKOPTS gcc@$GCC_VER %gcc@4.8.5
+  #spack install $SPACKOPTS gcc@$GCC_VER
+
+  # Use a gcc 8.5.0 "bootstrapped" gcc 8.5.0
+  spack install $SPACKOPTS gcc@$GCC_VER %gcc@$GCC_VER
+  spack compiler add `spack location -i gcc@$GCC_VER`/bin
+ 
+  cd $home
+}
+
+
+install_intel_oneapi () {
+
+  home=$PWD
+
+  . $SPACK_DIR/share/spack/setup-env.sh 
+  spack install $SPACKOPTS intel-oneapi-compilers@${INTEL_VER}
+
+  spack compiler add `spack location -i intel-oneapi-compilers`/compiler/latest/linux/bin/intel64
+  spack compiler add `spack location -i intel-oneapi-compilers`/compiler/latest/linux/bin
+
+  spack install $SPACKOPTS intel-oneapi-mpi@${INTEL_VER} %intel@${INTEL_VER}
+  spack install $SPACKOPTS intel-oneapi-mkl@${INTEL_VER} %intel@${INTEL_VER}
+
+  cd $home
+}
+
+
+install_netcdf () {
+
+  COMPILER=intel@${INTEL_VER}
+
+  home=$PWD
+
+  . $SPACK_DIR/share/spack/setup-env.sh
+
+  # use diffutils@3.7 - intel compiler fails with 3.8
+  # use m4@1.4.17     - intel compiler fails with newer versions
+
+  spack install $SPACKOPTS netcdf-fortran ^netcdf-c@4.8.0 ^hdf5@1.10.7~cxx+fortran+hl~ipo~java+shared+tools \
+     ^intel-oneapi-mpi@${INTEL_VER}%gcc@${GCC_VER} ^diffutils@3.7 ^m4@1.4.17 %${COMPILER}
+
+  cd $home
+}
+
+
+install_esmf () {
+
+  COMPILER=intel@${INTEL_VER}
+
+  home=$PWD
+
+  . $SPACK_DIR/share/spack/setup-env.sh
+
+  spack install $SPACKOPTS esmf%${COMPILER} ^intel-oneapi-mpi@${INTEL_VER}%gcc@${GCC_VER} ^diffutils@3.7 ^m4@1.4.17 \
+      ^hdf5/qfvg7gc ^netcdf-c/yynmjgt
+
+  cd $home
+}
 
 
 install_base_rpms () {
@@ -180,15 +285,13 @@ install_base_rpms () {
   tar -xvf $libstar
   rm $libstar
  
-  # GCC needs to be installed first
   rpmlist='
-    gcc-6.5.0-1.el7.x86_64.rpm
     hdf5-1.10.5-4.el7.x86_64.rpm
     netcdf-4.5-3.el7.x86_64.rpm
-    esmf-8.0.0-1.el7.x86_64.rpm
     produtil-1.0.18-2.el7.x86_64.rpm
+    esmf-8.0.0-1.el7.x86_64.rpm
   '
- 
+
   for file in $rpmlist
   do
     sudo yum -y install $file
@@ -205,6 +308,7 @@ install_extra_rpms () {
 
   home=$PWD
 
+  # tarball contents
   # bacio/v2.1.0         w3nco/v2.0.6
   # bufr/v11.0.2         libpng/1.5.30        wgrib2/2.0.8
   # g2/v3.1.0            sigio/v2.1.0
@@ -238,7 +342,7 @@ install_extra_rpms () {
   done
 
   # Force install newer libpng leaving the existing one intact
-  # this one will be used for our model builds
+  # this one will be used for our model builds via the module
   sudo rpm -v --install --force libpng-1.5.30-2.el7.x86_64.rpm  
 
   rm -Rf "$wrkdir"
@@ -256,15 +360,15 @@ install_python_modules_user () {
   . /usr/share/Modules/init/bash
   module load gcc
   sudo python3 -m pip install --upgrade pip
-  python3 -m pip install --user wheel
-  python3 -m pip install --user dask distributed
-  python3 -m pip install --user setuptools_rust  # needed for paramiko
-  python3 -m pip install --user paramiko   # needed for dask-ssh
-  python3 -m pip install --user prefect
-  python3 -m pip install --user boto3
+  python3 -m pip install --user --upgrade wheel
+  python3 -m pip install --user --upgrade dask
+  python3 -m pip install --user --upgrade distributed
+  python3 -m pip install --user --upgrade setuptools_rust  # needed for paramiko
+  python3 -m pip install --user --upgrade paramiko   # needed for dask-ssh
+  python3 -m pip install --user --upgrade prefect
+  python3 -m pip install --user --upgrade boto3
 
-
-  # Build and install the plotting module
+    # Build and install the plotting module
   # This will also install dependencies
   cd ../..
   # Should be in ~/Cloud-Sandbox/cloudflow
@@ -341,14 +445,71 @@ install_impi () {
 
   sudo ./aws_impi.sh install -check_efa 0
 
-  version=`cat /opt/intel/compilers_and_libraries/linux/mpi/intel64/modulefiles/mpi | grep " topdir" | awk '{print $3}' | awk -F_ '{print $4}'`
-
   sudo mkdir -p /usrx/modulefiles/mpi/intel
-  sudo cp -p /opt/intel/compilers_and_libraries/linux/mpi/intel64/modulefiles/mpi /usrx/modulefiles/mpi/intel/$version
+  
+  # The included module file does not work
+  # version=`cat /opt/intel/compilers_and_libraries/linux/mpi/intel64/modulefiles/mpi | grep " topdir" | awk '{print $3}' | awk -F_ '{print $4}'`
+  # sudo cp -p /opt/intel/compilers_and_libraries/linux/mpi/intel64/modulefiles/mpi /usrx/modulefiles/mpi/intel/$version
 
-  cd $home
+  cd /opt/intel/impi
+  version=`ls -1 | grep 20??\.??\.*`
+ 
+  sudo tee /usrx/modulefiles/mpi/intel/$version << EOF
+#%Module1.0#####################################################################
+#
+# Copyright 2003-2019 Intel Corporation.
+# 
+# This software and the related documents are Intel copyrighted materials, and
+# your use of them is governed by the express license under which they were
+# provided to you (License). Unless the License provides otherwise, you may
+# not use, modify, copy, publish, distribute, disclose or transmit this
+# software or the related documents without Intel's prior written permission.
+# 
+# This software and the related documents are provided as is, with no express
+# or implied warranties, other than those that are expressly stated in the
+# License.
+#
+##
+## Intel(R) MPI Library modulefile
+##
+
+proc ModulesHelp { } {
+        global dotversion
+        puts stderr " Intel(R) MPI Library"
 }
 
+module-whatis       "Sets up the Intel(R) MPI Library environment"
+
+set                 topdir                 /opt/intel/compilers_and_libraries
+
+setenv              I_MPI_ROOT             \$topdir/linux/mpi
+
+prepend-path        CLASSPATH              \$topdir/linux/mpi/intel64/lib/mpi.jar
+prepend-path        PATH                   \$topdir/linux/mpi/intel64/bin
+prepend-path        LD_LIBRARY_PATH        \$topdir/linux/mpi/intel64/lib/release:\$topdir/linux/mpi/intel64/lib
+prepend-path        MANPATH                \$topdir/linux/mpi/man
+
+if { [info exists ::env(I_MPI_OFI_LIBRARY_INTERNAL) ] } {
+    set i_mpi_ofi_library_internal $::env(I_MPI_OFI_LIBRARY_INTERNAL)
+} else {
+    set i_mpi_ofi_library_internal "yes"
+}
+
+switch -regexp -- \$i_mpi_ofi_library_internal {
+    0|no|off|disable {
+    }
+    default {
+        setenv              FI_PROVIDER_PATH       \$topdir/linux/mpi/intel64/libfabric/lib/prov
+
+        prepend-path        PATH                   \$topdir/linux/mpi/intel64/libfabric/bin
+        prepend-path        LD_LIBRARY_PATH        \$topdir/linux/mpi/intel64/libfabric/lib
+        prepend-path        LIBRARY_PATH           \$topdir/linux/mpi/intel64/libfabric/lib
+    }
+}
+EOF
+  cd $home
+}
+#####################################################################
 
 
 # Personal stuff here
